+++ src/Symfony/Bundle/Resources/config/normalizers.php
@@ -4,6 +4,7 @@ namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
 use Xen3r0\JiraApiClient\Serializer\Normalizer\Issue\CommentNormalizer;
 use Xen3r0\JiraApiClient\Serializer\Normalizer\Issue\FieldsNormalizer;
+use Xen3r0\JiraApiClient\Serializer\Normalizer\Issue\WorklogNormalizer;
 
 return static function (ContainerConfigurator $container) {
     $container->services()
@@ -12,6 +13,10 @@ return static function (ContainerConfigurator $container) {
             ->tag('serializer.normalizer')
 
         ->set(FieldsNormalizer::class)
+            ->args([service('serializer.normalizer.object')])
+            ->tag('serializer.normalizer')
+
+        ->set(WorklogNormalizer::class)
             ->args([service('serializer.normalizer.object')])
             ->tag('serializer.normalizer');
 };